# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ava Project API",
    "description": "Local AI assistant backend. Voice pipeline, LLM agent, knowledge base, plugins, and more.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/mcp/servers": {
      "get": {
        "tags": [
          "mcp"
        ],
        "summary": "List Servers",
        "description": "List all registered MCP servers with their current status.",
        "operationId": "list_servers_api_mcp_servers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response List Servers Api Mcp Servers Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "mcp"
        ],
        "summary": "Register Server",
        "description": "Register a new MCP server.",
        "operationId": "register_server_api_mcp_servers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpServerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Register Server Api Mcp Servers Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mcp/servers/{name}": {
      "put": {
        "tags": [
          "mcp"
        ],
        "summary": "Update Server",
        "description": "Update an existing MCP server config.",
        "operationId": "update_server_api_mcp_servers__name__put",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "mcp"
        ],
        "summary": "Delete Server",
        "description": "Unregister and stop an MCP server.",
        "operationId": "delete_server_api_mcp_servers__name__delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mcp/servers/{name}/start": {
      "post": {
        "tags": [
          "mcp"
        ],
        "summary": "Start Server",
        "description": "Start a registered MCP server (resolves secrets from vault).",
        "operationId": "start_server_api_mcp_servers__name__start_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Start Server Api Mcp Servers  Name  Start Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mcp/servers/{name}/stop": {
      "post": {
        "tags": [
          "mcp"
        ],
        "summary": "Stop Server",
        "description": "Stop a running MCP server.",
        "operationId": "stop_server_api_mcp_servers__name__stop_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mcp/servers/{name}/tools": {
      "get": {
        "tags": [
          "mcp"
        ],
        "summary": "List Server Tools",
        "description": "List all tools exposed by an MCP server.\n\nIncludes error details if tool discovery failed, and the timestamp of\nthe last successful discovery.",
        "operationId": "list_server_tools_api_mcp_servers__name__tools_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Server Tools Api Mcp Servers  Name  Tools Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/mcp/servers/{name}/logs": {
      "get": {
        "tags": [
          "mcp"
        ],
        "summary": "Server Logs",
        "description": "Get recent stderr logs from an MCP server.",
        "operationId": "server_logs_api_mcp_servers__name__logs_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Server Logs Api Mcp Servers  Name  Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mcp/servers/{name}/oauth/start": {
      "get": {
        "tags": [
          "mcp"
        ],
        "summary": "Oauth Start",
        "description": "Return the OAuth authorization URL for this MCP server.\n\nThe server config must have oauth_url, oauth_client_id, and optionally\noauth_scopes set.  The redirect_uri is always pointed back to the\nbackend callback endpoint so the token exchange can happen server-side.",
        "operationId": "oauth_start_api_mcp_servers__name__oauth_start_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Oauth Start Api Mcp Servers  Name  Oauth Start Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mcp/servers/{name}/oauth/callback": {
      "get": {
        "tags": [
          "mcp"
        ],
        "summary": "Oauth Callback",
        "description": "Exchange the OAuth authorization code for an access token.\n\nStores the token as an encrypted secret named `{NAME}_OAUTH_TOKEN`\nso that the MCP server env can reference it as `${NAME_OAUTH_TOKEN}`.",
        "operationId": "oauth_callback_api_mcp_servers__name__oauth_callback_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Authorization code from OAuth provider",
              "title": "Code"
            },
            "description": "Authorization code from OAuth provider"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Oauth Callback Api Mcp Servers  Name  Oauth Callback Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "McpServerCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "command": {
            "type": "string",
            "title": "Command"
          },
          "args": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Args"
          },
          "env": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Env"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "auto_start": {
            "type": "boolean",
            "title": "Auto Start",
            "default": false
          },
          "oauth_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oauth Url"
          },
          "oauth_client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oauth Client Id"
          },
          "oauth_scopes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oauth Scopes"
          },
          "oauth_client_secret_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oauth Client Secret Ref"
          }
        },
        "type": "object",
        "required": [
          "name",
          "command"
        ],
        "title": "McpServerCreate"
      },
...
      "OkResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "title": "OkResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}